ride-backend/drift-ingestion-service/requirements.txt
apscheduler>=3.10.1
redis>=4.5.1
requests>=2.31.0
pandas>=2.0.0 
boto3>=1.35.1
fastapi
uvicorn[standard]
numpy
websockets
python-multipart
slowapi
h3==3.7.7
python-dotenv==0.19.0
gtfs-realtime-bindings>=0.0.7
google-cloud-secret-manager>=2.16.0
